requirements.txt
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.3,<5.0.0
python-dotenv>=1.0.1,<2.0.0
schedule>=1.2.1,<2.0.0
pytz>=2024.1,<2025.0
